summarizer_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
openai>=0.27.0
python-dotenv>=0.21.0
pytest>=7.0.0
